packages/voiceshared/pyproject.toml
[project]
name = "voiceshared"
version = "0.1.0"
description = "Shared components for Azure Voice Live Support Agent services"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "azure-ai-voicelive>=0.2.0",
    "azure-identity>=1.19.0",
    "aiohttp>=3.9.0",
    "openai>=1.58.0",
    "python-json-logger>=3.2.0",
    # Microsoft Agent Framework (meta-package installs core + provider sub-packages
    # such as agent_framework.openai and agent_framework.foundry).
    "agent-framework>=1.0,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
